requirements.txt
# Core
fastapi
uvicorn
streamlit
python-dotenv
pydantic

# LangChain + LangGraph
langchain
langchain-core
langchain-groq
langgraph

# Database
mysql-connector-python
sqlalchemy

# Vector Store
faiss-cpu
sentence-transformers

# HTTP & APIs
requests
httpx
twilio

# Scheduling
apscheduler

# Utilities
loguru
python-multipart
